requirements.txt
argcomplete==1.12.2
boto==2.49.0
boto3==1.16.25
botocore==1.19.25
certifi==2020.11.8
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
durationpy==0.5
eib-aws-utils==1.0.5
Flask==1.1.2
future==0.18.2
hjson==3.0.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
MarkupSafe==1.1.1
pip-tools==5.4.0
placebo==0.9.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-json-logger==2.0.1
python-slugify==4.0.1
PyYAML==5.3.1
requests==2.25.0
requests-toolbelt==0.9.1
s3transfer==0.3.3
six==1.15.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.54.0
troposphere==2.6.3
Unidecode==1.1.1
urllib3==1.26.2
webexteamssdk==1.6
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.52.0
